Fixing the update target when using PLUGIN_DIR [devel]
The update of the dgb repository was not working because of using the incorrect
work directory. A specific target needs to be used for that.

Now the new target "update_all" updates, od, its external, dgb and its externals in
that specific order.
cmake does not have tools for hiding the temporary targets update and update_dgb,
which thus still appear in visual studio.