Added cmake support for OpenSSL [devel]
The following two macro allow any OpendTect module to link against
openssl and/or crypto library, and require the installation of the
corresponding libraries:
OD_SETUP_OPENSSL
OD_SETUP_CRYPTO

Used to install libopenssl.so on Linux, and update the setup
for the plugin dGBPDF3D, that needs to link against libcrypto.so and
install it.

Supports either the system libraries, or the libraries installed using the
Qt Maintenance tool.