Fixes in cmake setup for openssl [devel]
The test program test_networkaccess clearly shows that the openssl
runtime library (so/dll) needs to be available on all platforms.
This commit adds the setup for windows, and makes the test program work
on Linux too, although we will need to remove the introduced dependency to
openssl for the Network module.