Fix 2D SEGY import
Summary:
Various changes to make 2D SEGY import to CBVS  work in trunk. Note I have replaced the
orignal multi-threaded soltuion with a single thread for input and output because
there is a bug in the threaded code that stops import of 2D data that I can't find.
Another issue is that OpendTect has to be restarted to display any data imported. The
2D line geometry and trace data is accessible from their respective managers after load
but adding the newly imported data to the tree doesn't work
- Merge branch 'master' into OD2853
- Fix batch 2d segy import to CBVS
- Fix interactive 2D SEGY import to CBVS file
- Clean up code

Test Plan:
Tested on Linux. 2D SEGY import to CBVS both interactive and batch is now working for
single and multiple lines. Note comment in summary regarding restart to display
imported data.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13542