Attempt to restore ABI compatibility [devel]
Summary: When the body of a function is moved from CC to header file, there are chances (in this particular case, almost 100%) that the compiler may inline it. In that case, the previously linked binaries will no longer find the definition in the shared library. The other way around (moving from header to CC file) is usually safe, as long as the previous (inlined) definition is still acceptable.

Test Plan: Compiles. I hope it will solve the runtime error in libMistie.so

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13549