Upgrade of cmake setup to export the module dependencies [devel]
FindOpendTect.cmake did not contain enough information about the
external dependencies. Now exporting the list of external libraries
for each module, and the corresponding module setup.
Includes a major rewrite of ODQtUtils.cmake, to clearly separate the
external per Qt component (Core/Gui/WebEngineWidgets).
Also adopting the same strategy for uiwebengine as in od6.6: it gets
excluded from uiBase, to be included in a dedicated plugin.