Using the ninja generator for building the externals [devel]
When available, using the ninja generator speeds up the
building of the externals, as it is automatically using
all cpus.
Directly usable on windows with MSVC2019, which ships ninja
internally.
Requires to set an alias to ninja or have it in your path on Unix.
Fallsback automatically to Unix/MSVC generators if ninja is not
available.