Fixed two tests programs [devel]
test_file and test_bendpoints2coords need a file input,
but they are not (cannot) be batch programs, as they are located
in Basic/Algo.
Hence they must be able to run without having to provide a parameter file argument.
Made that clearer by changing the file extension.
Also fixed an error when building within the source directory
(then __FILE__ is relative to it, and it not a fullpath)