Other fix for running python scripts on Unix [devel]
When run with virtual environment activation, the commands
to run are placed in a script file. In that script file the
arguments must be single quoted for the command to run.
However when the script flag -c is used, the single quotes do
not work, and need to be replaced with double quotes.