Fixed checking of debugging symbols
One only needs perform checks on the first output line,
thus it is more effecient to first extract the first line
of the output, before splitting it into a list from the spaces.
It also avoids a cmake warning/error that multiple contiguous
spaces will create empty elements in the output list.