Removed a mis-leading notifier for uiDialog [devel]
Summary:
applyOKCB, just like acceptOKCB and rejectOKCB is no longer
a virtual function, and the corresponding notifiers for acceptOK and
rejectOK had already been removed.
The correct (new) way to implement an action when the apply button is
pressed is thus to re-implement the applyOK() function, like for
rejectOK and acceptOK. Hence the Notifier applyPushed should not be used
and should be removed.

We should take that approach in the main branch too, where all three
functions need to be migrated.

Test Plan: tested from MachineLearning plugin

Reviewers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13605