Fixed crash when adding a Stratigraphy Unit
Summary:
the crash occur when the local object tmpun of type LeavedUnitRef
is removed at the end of the function uiStratRefTree::insertSubUnit when it
gets out of scope.
The only fix I could image is in the implementation of Strat::UnitRef::notifChange.
Fixes OD-2910

Test Plan: Successfully added a unit. No crash

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13609