Activate: check for presence in bin and condabin
Summary: When anaconda3 is installed there can be a bin directory on Windows, therefore we will be checking for the presence of the activation script in both bin and condabin folder and not stop the checking because there is a bin folder.

Test Plan:
Tested on Windows 10, when activate.bat is in bin or condabin folder is is found. Not tested right now, but on Linux it should find activate script in bin folder and if not then it will also check the condabin folder for just in case.

# Tip: You can add a project as a subscriber or reviewer by writing
# "#projectname" in the appropriate field.

# Tip: Create a dependency between revisions by writing "Depends on D123" in
# your summary.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13615