Add granularity to loading of well logs in the well manager
Summary:
Fixes OD-2826
The is a less ambitious approach to this issue where I have not implemented
use of smartpointers to automatically manage memory. I've attempted to ensure that
this will work with existing well related code in the odmain and dgbmain branches by
ensuring the previous well API is intact.
To get full benefit from the changes a new field is required in the well log headers containing
the log value range. The implementation will work if this field is not present but in that
case the log will be loaded to memory so it can be scanned to set the value range.
The well writer has been modified to automatically add this log value range field to
the log file header during output.

Test Plan: Tested on Linux.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: prajjaval

Differential Revision: https://phabricator.dgbes.com:8443/D13619