Changle default input units in well log calcu user formula
Summary:
Fixes OD-2953 Well Log mathematics anomalies
Previously the default setting for a user formula was to convert the input log to
SI units before calculation. This change makes the default to keep the input in its
original units. This should be more natural for the user because any constants entered
in the formula will probably be defined in the units of the input log.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13624