Use parallel processing for horizon shift tool
Summary:
Fixes OD-2959 Horizon shift is extremely slow
I've atempted to speed this up by using parallel processing for both filling the
datapointset locations and extracting the attributes. This seems to help, however,
this is a very slow piece of code. Probably would be faster to generate a volume around the
horizon and sample that then using a datapointset.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13630