conda envs and IDE can be used via .conda txt
Summary: The user can now put the custom conda environments in different path. These paths are saved into userdir\.conda\environments.txt. The IDE can also be used via the custom path.

Test Plan: Tested on Windows 10: loading of envs via txt file, loading of envs when txt file is not present, testing launching of Spyder IDE. Please note that when user has in user directory the file .condarc it will use the envs from the alternative path when launching the terminal even when setting to the basedir to OpendTect Python dir. This is probably a normal behavior because conda will read this file and will find the environment in the alternative path. This will need additional testing on Linux and also with the ML plugin, but for that the added code needs to be ported to od6.7

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D13634