Force regeneration of cmake project when the externals have changed [devel]
When the external have changes, they need to be reconfigured and rebuild.
This only happens if one reconfigures the OpendTect project, not be running
a project target.
By updating the timestamp of the OpendTect CMakeCache.txt file, this
regenerating should by done automatically as cmake monitors the
access time of this file when running any target.