Refinements to well display property dialog
Summary:
Changes to fix OD-2969 depends on D13644
1. Seismic style display is centered on the well track for the Center Log display, Repeat UI field
must be odd - the spinbox settings changed to enforce this
2.Reading display settings and well display issues fixed
3.Added a welldispSettings file to the data folder to define default display settings - this
may need to be tweaked. Also added a check in the display properties dialog that the colors
of the log and track have good contrast with the scene background. If not the colors are changed - the
user can go back to the original colors by using the Reset button.
4.Center Log has been made the default tab and the requested default style changes have been applied
5.Added Reset button that reads the display settings for the current well from disk and overwrites any
unsaved changes in the dialog box.
6.Added "Apply Current to all wells" button that applies the contents of the visible tab to all
loaded wells. In the case of logs the change is only applied if the log exists in the well.
Changed the "Apply to all" button to "Apply All to all wells". Log display settings only change
if the log exists in a well.
Added a "Reset All" button that overwrites any changes to well display settings with the settings stored
on disk for all wells

Test Plan: Tested on Linux

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13645