Expand log display vertically over window extent
Summary:
Fixes OD-2994 Well Tools display window: fixed size too small
To get a working solution I had to settle on using a fixed width for the
log panels to get the horizontal scroll bar to behave. The vertical size
of the log displays is now responsive to the window size.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13672