Fixed breakpad initialization [devel]
The initialization of breakpad was done both in initbasic and inituibase,
and depended on a flag defined in Basic but not accessible to uiBase.
Made the class CrashDumper always accessible, and move the few breakpad specific
implementation under ifdef within the function implementations.
Also cleaned up the command line for windows, and its parsing in the process_dumpfile.cmd
script.

Tested that a triggered crash is send using od_uiReportIssue, and can be
converted to a readable callstack.