Added full support of external python environments [devel]
Fixed a small bug where the base environment was not usable.
Made it work with miniconda, where the name of the environment needs to
be the full path to the external environment.
Also works if that full path contains a space.
Finally, the custom option will now list all virtual environments when pointing
to the OpendTect-installed miniconda installation, and not only the environments
with a priority setting.