Added support of external python virtual environments
The user can now put the custom conda environments in different path.
These paths are saved into userdir\.conda\environments.txt. The IDE can also be used via the custom path.

Tested with both Anaconda and Miniconda, where the name of the environment needs to
be the full path to the external environment.
Also works if that full path contains a space.
Finally, the custom option will now list all virtual environments when pointing
to the OpendTect-installed Miniconda installation, and not only the environments
with a priority setting.