Various cmake changes for Qt and OpenSSL [devel]
Summary:
Added installation of test programs, so we can transfer them to any machine after a release is made, to test some functionality.
Resolved the full path to libssl, libcrypto library because the implementation needs the target filename.

Qt compliance:
Qt plugins should be installed outside our Debug/Release folder, at the root of the Qt prefix.
[Linux only] QtWebEngineProcess should go in a libexec directory, at the root of the Qt prefix.
From Qt deployment tests:
Removed the plugin platformthemes, not required by uiBase or uiOSG
Added the plugins bearer and position with the WebEngine.
This later does not need the executable qwebengine_convert_dict, thus removed.
Added the library d3dcompiler_47.dll for windows.

Linux only (and Mac?):
By creating a soft link 'lib' that points to either 'Debug' or 'Release' folder, all our executables now can start without any script. Changing the keyword 'Libraries' from lib to Debug/Release in qt.conf does not work.

Removed CMAKE_INSTALL_PREFIX where applicable, as it it preferred to provide a relative path to it

Removed the path to the OpendTect executables from both .init_dtect and run_installer, as this is hard-coded with the $ORIGIN keyword by cmake when all targets are installed.
Fixed a minor error on some linux machines with the command grep -m 1 GLIBC_2.17

Added the new folder OpenSSL to LD_LIBRARY_PATH in .init_dtect, but only if the library libdGBPDF3D.so is installed, and only if the system does not have such library (replaced previous compatibility check.

RuntimeLibLoader can now accommodate a sub directory.

Linux only:
The dynamic library loading in Network will now skip the loading of the OpenSSL libraries provided with the release if the system libraries are new enough, by looking for a specific symbol.