Fix display of centerlog
Summary:
Fixes OD-3025 Center log well display setting not working
I've ported across some more changes from the main branch required
for the centerlog display.
I haven't fully tested it but the obvious things seem to be working.
The issue I was having was resolved by using a pointer as opposed
to the reference to the pointer to make the calls into osgGeo. Although this
works ok in the main branch it mustn't be compatible with pointers coming from
a HiddenParam.
The display properties dialog in 6.6 is behaving differently to the main branch.
I'll move onto that next.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13713