Fixed a warning when linking a WIN32 executable with breakpad on [devel]
cmake always links targets dependent on QtCore against qmain.lib, but Qt
does not provide a pdb file for this library. As a result the warning 4099
is issued when linking a WIN32 executable built in release mode with breakpad,
as we turn on the generating of debug symbols for these executable too.
Warning is gone.