Improved safety of applicationClosing [devel]
Very few classes should be able to call DBM().applClosing(). They are now
listed as friend classes, and the corresponding function is made private.

For the few application that do not have either a QCoreApplication, or QGuiApplication,
the new class DBMManager ensures that applClosing is still triggered when the
program closes (provided that the General module was loaded/initialized).