Improved safety of applicationClosing [devel]
Very few classes should be able to call IOM().applClosing(). They are now
listed as friend classes, and the corresponding function is made private.

For the few application that do not have either a QCoreApplication, or QGuiApplication,
the new class IOMManager ensures that applClosing is still triggered when the
program closes (provided that the General module was loaded/initialized).