Delayed the AppLocker tests to after all plugins are loaded [devel]
File::initTempDir cannot stay in the initialization of Basic, because it
needs to become conditional: No AppLocker tests should be done when using
OpendTect-PRO ( conditioned to the initialization of dGBCommon that modifies
OD::PythA() ), or in the absence of python modules shipped with the application
(like for the installer).

This fix is good enough for most OpendTect applications, but not yet for both
od_DBMan and od_WellMan that should also skip these tests.