Fix 3d display of wells after using Well Table
Summary:
Fixes OD-3059 Wells do not display properly in 3D scene if displayed after opening Well Table
Currently there is no way to determine if a well has had its display parameters loaded from storage or
is just using defaults. The Well Table creates the well data object but doesn't load the display parameters.
If a well is loaded by the Well Table and then displayed in the 3D scene it will be shownn using the
defaults. This change forces a reload of the display paraneters when the well is added to the
3D scene.
Also fixes a potential issue where only a partial set of logs is loaded in a well data object. Now
Logs and LogInfos will only be set in the loadState if the in memory log set contains all the logs
stored on disk.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13758