Temporary fix to avoid crashes when closing the main application [devel]
All children windows are deleted by Qt when the body of the top level uiMainWin
gets deleted. However the uiMain object, which triggers DBM().applClosing in its
destructor, is deleted much later.
As a result all classes that try to delete non-modal windows in their destructor,
or on applicationClosing, will trigger a crash because of double deletion.
This fix will allow legacy code to delete the non-modal windows inside applicationClosing,
before the body of the top level uiMainWin gets deleted.