Fixed closing of non-modal windows when the main app closes [devel]
Introducing the global function closeAndZeroPtr(), to be used in class which
keep a pointer to a non-modal window. This function should be used to close
the window at request.
Qt deletes all uiMainWin and uiDialog windows automatically, thus removed
all corresponding deletion from classes desctructors.

Non-modal windows without a parent need to listen to windowClosing of the
topLevel application in order to automatically close at the same time.