Fixed corrupted export of 2D horizons [devel]
Summary:
Looks like in a statement like
    controlstr.add( prov.xy() ).add( prov.trcnr() ).add( prov.spnr() ).add( prov.string() );
MSVC evaluates all calls to the prov object before executing the add( ) calls. As a result,
since a static string is returned and overwritten, all added strings corresponded to the first output:
prov.xy() and the output was corrupted.
Thus added a static string in each function from the FormatProvider class, which would be sufficient,
provided that the same class member function is never called twice on the same line. Hence FormatProvider::xy()
returns the format string for both X and Y now.

Fixes OD-3068.