[cmake] Removing update_all target on windows
Because this cmake target is dependent on others, the dependent
targets are added by cmake to the ALL target, and always build when building the
solution on Windows, which is not the intention. These targets should only be build
at explicit request.
Thus disabled that target on Windows for now.