Implement on-demand log loading for 3d display
Summary:
This seems to be a viable solution to defer loading of logs until they are displayed.
My debug build of the main branch is still noticeably slower than either 6.4 or 6.6.3rc6 for
loading the ML force data (well track display only). Not sure if this could just be the debug
build overhead. I'll do a Release build tomorrow for a more direct comparison.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13798