Improved OpendTect isolation script on Linux [devel]
The script was relying on the presence of the environment
variables OD_INTERNAL_CLEANPATH and OD_SYSTEM_LIBRARY_PATH to
reset the environment before sending a command.
These variables may not be set, for instance they are not set
by the installer which does not execute the init_dtect script.

The new implementation uses DTECT_APPL if set to remove any path
in $PATH and $LD_LIBRARY_PATH that would start with $DTECT_APPL.

When DTECT_APPL is not set, it gets set to the installation
containing the script od_external.sh

Should fix the dynamic linking error when the installer starts on
Linux, which tries to use the Qt libraries of the OpendTect installation
instead of using its own libraries.