Further required improvements for od_external.sh [devel]
Previous cleanup of PATH/LD_LIBRARY_PATH was still not good enough.
Now removing from these variables any folder that contains either
the OpendTect Basic library or the Qt core library.
This will guarantee that there won't be any conflict between the
application to be launched and the calling environment, because
of mistmaching OpendTect and/or Qt library versions.