Further required improvements for od_external.sh [devel]
Summary:
Previous cleanup of PATH/LD_LIBRARY_PATH was still not good enough.
Now removing from these variables any folder that contains either
the OpendTect Basic library or the Qt core library.
This will guarantee that there won't be any conflict between the
application to be launched and the calling environment, because
of mistmaching OpendTect and/or Qt library versions.

Test Plan:
I have manually set LD_LIBRARY_PATH such that it contains the
full path to the OpendTect libraries of an old installer, and the full path
to the libraries of an old Qt development installation. Thereafter it was
no longer possible to start successfully an OpendTect GUI application, unless
calling od_external.sh followed by the OpendTect application name.
Also tested the same command line with an empty or unset LD_LIBRARY_PATH

Reviewers: raman, nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13810