Avoid database initialization for application which do not need it [dev]
Summary:
Storing the argument given to SetProgramArgs in order to
avoid database-related initialization in applications like the installer,
the image viewer, progress viewer, ...

Test Plan:
Launched od_DispMsg and od_main with a breakpoind in SurveyInfo, Survey::Geometry3D
and DBMan. Checked that it was only hit for od_main, and not od_DispMsg.

Reviewers: nanne, raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13819