Fixed incorrect angle gathers in SynthRock
Summary:
AngleComputer::computeAngleData did not use the
virtual function curRayTracer when raytracing needed to be done.
As a result the first run was filling the raytracing_ class member of
the base class, instead of using a RayTracer1D object stored in the
ModelBasedAngleComputer.
As a result all calls to AngleComputer::fillAndInterpData
were using the RayTrace1D object from the first sequence.

Fixes OD-3105

Test Plan:
using the model description attached to the bug. This
one shows clearly that the angle gather get cropped at a constant
TWT, instead of following the dip of the model.
Also vieweing the angle gathers in the gather viewers now shows
that the angle gather strectch conformably with the model.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13823