Fixed failure to restore Z Range in Machine Learning extraction
Summary:
An IOPar saved by Survey::FullSubSel could not be used to
initialize the Pos::Provider in uiIo because of incompatible keys.
With this fix the IOPar of both classes are compatible again.

Also added several operators that will be needed to fix OD-3107,
especially the == operator in SeisSel.

Test Plan:
Saved Machine Learning data extraction parameters in
an Img2Img workflow. The user-interface did not show the Z range
stored in the saved file, now it does, either by opening a
parameters file from a legacy version, or with the current code.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13825