Fix for importing horizons when X/Y and Inl/Crl have similar value ranges
Summary: In the SEAM survey, X/Y values are rather small, making them comparable to Inl/Crl values. The scanner can easily get fooled by this. Now the scanner first tries the user-specified position type (Coord or BinID) and only when this option fails it tries the alternative.

Test Plan: Successfully imported a horizon in SEAM survey and matched it with the original horizon.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13830