Fix Well Table: Well log availability issue
Summary:
Fixes OD-3114 Well Table: Well log availability is showing in red color for available logs too
This was occuring because old format well log files do not have the Dah range stored
in the file header.

The change ensures the Dah and log value range is set in the in-memory well storage. This
requires the entire log file to be read even if only the info is requested.

For the future use, where absent, the ranges are automatically saved to the well log file header.
While this involves writing out the log this is a one time event.

Test Plan:
Tested on Linux:
- tested on project with mixture of log files with and without Dah and log ranges in the file headers
- all logs correctly displayed in the well table
- confirmed by viewing file modification stamps that only files without Dah and log ranges were rewritten

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13836