Updated HDF5 interface to better support resizable datasets [devel]
SUMMARY
Current interface had only support for setting a global chunk size
for all write operations. This is not good enough.
Chunking is dataset specific, together with resize (crop/extend) operations, hence
it should not be set at the writer level. By adding it to the DataSetKey object,
the corresponding properties can be passed to all write functions, and become
natively dataset specific.
In general setting the chunk size to a hard value should be avoided, as it will generate
overhead in disk/memory usage and I/O speed. What is really needed, and added with
this code, is the capability to specify at least one axis that needs to stay resizable,
most often it will be the first axis. That would be the typical usage to be able to
add/remove rows of a 2D table, or add/remove examples from a Machine Learning dataset.
All other dimensions remain fixed to the input ArrayND sizes.

TEST PLAN
used the test program. Updated the usage of setting the chunk size, but
also made sure that some datasets get written without any chunking.
The 5D dataset "BigData" had already an implementation of cropping, added an
implementation of append. This dataset is now set using the typical usage that only
the first axis is resizable.
Since the chunking/resizability is now correctly implemented, the resulting file
size of this test program is divided by 3.

Needed to solve OD-3107 / OD-3110.