Made the well display properties dialog non-modal again
Summary:
For OD-3115. Looks like it changes the ABI, but actually I restored
the ABI as compared to the public 6.6.2 release, while keeping the
changes (new constructors) to the bare minimum.
Main change in this revision is that I had to create a fake save button
since I cannot reimplement the acceptOK function of the dialog. We never
want anyway to close the dialog after pressing this save button.

Test Plan:
Loaded in the scene a well without/with saved display properties.
Checked that the correct software defaults are applied in the first case,
and that the interface displays the saved settings in the second case.
Used the various buttons and checked the behaviour to see if it matches the
intention.

For now I have added an explicit reload of the well data from database
when a well gets added to the scene. Not ideal, but there is no other way
at the moment to force refresh a well data without restarting the application.
Will need more work to handle that.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13846