Implemented resolution setting for 3D well log displays
Summary:
Number of samples in log displays was limited to 2000 samples. Implemented existing virtual functions.
Change in uiVisPartServer was needed to keep the Resolution menu. Setting the id to -1 caused it to disappear from the Display menu.

JIRA: OD-3125

Test Plan: Added wells in a couple of surveys. Default resolution seems OK. Resolution can be changed for wells individually.

Reviewers: raman, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13853