Fixed incorrect default for dynamic name size [devel]
SUMMARY
This settings, which was available in the well log display
properties for the track and markers, has never been saved in the
dataset settings, up to and including the public release 6.6.2.
The UI default was true.
However the non-ui default is false. Hence in order to correctly
handle the legacy IOPars, where the key is missing, I am setting
the class member to true when reading a compatible IOPar, when the
key is missing, otherwise the saved font size will make the label
most often print at the wrong size.

TEST PLAN
Displayed a legacy well, checked that the properties
in the properties dialog match the content of the file, and that
the "dynamic" option is switched on when missing.