Fixed component selection when exporting 2D data to SEG-Y
Summary:
The component field was populated correctly for 3D data, but not for 2D.
Now using SeisIOObjInfo to get this information, which works for both 2D and 3D.

Jira: OD-1758

Test Plan:
Calculated specdecomp on a 2D line. Then exported only the 50Hz component, and
imported again. Confirmed in the 3D scene that original 50Hz component is the
same as the imported one.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D13890