Several fixes in well display/loading
Summary: see OD-3140.

Test Plan:
Made sure the following workflows run as expected:
1- Loading wells in the tree without stored display properties does not show/create osg markers and
their text label, and it does not even read the markers from the database
2- Loading wells in the tree with display properties should honour them. However most
often these display properties contain an empty list of markers to be displayed. In that case
the application must behave like in -1-.
3- Propagation of saved display properties from data to users folder should forward only meaningful
properties, for instance no log display properties if no log was selected in the first place
4- "Apply Current to all wells" button needs to use both the list of selected and unselected markers
to effeciently propage that selection to the other wells, where the marker list can be quite difference,
with little or not overlap.