Fixed two serious errors in the LogTools [devel]
The GUI for the gate of the grubbs filter was set to a size of 0,
causing a memory leak up to a crash.
The copy operator of both track and time2depth models did not copy
all class members of the DahObjh class, causing (at least) a failure to
apply a FFT-based filter, because the track dahrange_ was not copied,
hence it was empty in the LogSampler.