Fixed saving of default Stratigraphy model (and 2 more fixes).
When changing projects, the closeOK fuction is called. But at that time
SI() will already point to the new survey, and thus save the default in the
wrong survey.
Now the default is set/saved when opening a model.

2nd fix: Improved look and feel. Lines between categories are now grey.
Boundaries of units black. Set z levels correctly.
Text in empty model was hardly readable and wrong. Changed to red font
and made it 'Right-click'.
Fixed alignment of zome text. It was too close to model top border.

3rd fix: removed detachAllNotifiers() from setSensitive call. Does not
belong there.