Added patch version in IOPar header
Summary:
Just pushed this change into the main branch. The reasoning is that sometimes
we need to know in which patch version the IOPar was saved. This makes it
easier to make changes in a parameter file and keep support for a previously
saved file

Test Plan:
Started a steeringcube process and confirmed the full version was stored in the
parameter file.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13990