added IOMan::isPresent(const MultiID&) overload [devel]
Summary: Calls to this function were made from the ML plugin, but it went to the other overload returning always false. So, all the input fields in the seis apply window were filled with the default cube.

Test Plan: The apply window in ML now shows the correct selection corresponding to each input.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13998