Fix for jupyter kernel.json + more fixes
Summary:
This fixes the changing of origpath in kernel.json which is needed for Jupyter to start. Also removed old directories from the list and another important fix makes the
script check the directories inside environments instead of always checking in the base environment, e.g. it would replace inside ssl/misc in environments because this directory is
present in base environment, but if directory is not present in base environment it did not do this. Will need further improvements when we create new Python packages.

Test Plan:
tested by executing the setup.Python script on dgb30, Ubuntu 20.04. Will need further testing also with the online and offline installer when we create new Python
packages.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D14008