Fixed pErrMsg because of uiPluginSel dialog [devel]
Summary:
Fixes this message while launching od_main:
(PE) N8CmdDrive11WindowStackE | src/uiCmdDriver/cmddriverbasics.cc:743 >> CmdDriver found broken window branch <<

Very likely because the dialog is created, but not deleted, before the real main window of
application is displayed for the first time. This ensures it is deleted before the call odmain->go(),
and before the plugins are loaded.

Also removed the call to showAlwaysOnTop(). It does not make sense, as the uiPluginSel dialog
is never shown at the same time as the uiODMain main window. On the other hand it had negative
side effects at least on Linux, as it would stay above any other OS window, even when moved or
raised after the uiPluginSel dialog gets on screen

Test Plan:
pErrMsg is gone. Now I can move a terminal/browser window above the uiPluginSel
dialog after it is displayed on screen.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14057