Fixed launching the installer from the OpendTect manu using V6.4.5.
Summary: Installer was looking wrong Qt libraries adnd was failed to launch if user installed 6.6.0 and 6.4.0 installation in the OpendTect installation directory. Mde changes in the run_installer script , checking the OpendTect folder name and setting LD_LIBRARY_PATH to empty in the case of 6.4.0 and 6.2.0

Test Plan: Launched Installer successfully from OpendTect version 6.4.5 and 6.6.4rc1.

Reviewers: arnaud, raman

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14061