Fixed incorrect parsing of drive letter file paths [devel]
A string like 'E:' was not parsed correctly into a FilePath:
'E:' was added as the first level, instead of placing 'E' as the prefix,
hence checking the number of levels did not return the expected value.

Also listing the files at the root of a windows drive returned and empty
list, because Qt expects a string like 'E:\' when constructing a QDir.