Closing the splash screen before showing the uiODMain window [devel]
On windows, keeping the splash screen alive while the scene is rendered causes
issues in the OSG event loop: gets blocked, the screen is corrupted unless
switching to another project.
Thus instead of calling the finish function of the splash screen, removing it
completely before the go function of uiODMain is called.