Added (back) the startup splash screen
Summary:
Was silently commented in:
4f264ad#diff-69a0803252c841dd0c02f5e140c02a7b3c9394972c3784163967d29f16fba33c

and then removed in:
1c42133#diff-69a0803252c841dd0c02f5e140c02a7b3c9394972c3784163967d29f16fba33c

However it does provide a nice picture that let's the user know
that something is happening while the plugins are loading, and
the scene is initialized.

Test Plan:
On windows, keeping the splash screen alive while the scene is rendered causes
issues in the OSG event loop: gets blocked, the screen is corrupted unless
switching to another project.
Thus instead of calling the finish function of the splash screen, removing it
completely before the go function of uiODMain is called.
Splash screen is correctly displayed between the plugin selection window and
uiODMain.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14100